refactor: factor out markdown processing utils
